+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Forward AtaDevice `target_temperature_step` calls to
 `temperature_increment`.
 - Rename ATW zone `state` to `status`.
+- Rename ATW `state` to `status`.
+- Return heat statuses to `heat_zones` and `heat_water`
 
 ### Fixed
 - Fix `get_devices` type hints.